+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Changelog
 
+## 0.0.3
+
+- added support for formdata and files
+- added method set_files in PostRequest class for overide files
+- added method set_data in PostRequest class for overide text parameters
+- using <https://httpbin.org> for tests
+- updated readme
+
+## 0.0.2
+
+- updated setup.py
+- fix forgotten debug print
+
 ## 0.0.1
 
-- fork from from https://github.com/k3rn3l-p4n1c/postpython
+- fork from from <https://github.com/k3rn3l-p4n1c/postpython>
 - update core to import Postman collections V2.1
 - Added loader for Postman enviroments
 - Change enviroment variables case Insensitive to case sensitive